requirements.txt
configparser==6.0.0
python-telegram-bot==13.7
requests==2.31.0
telegram==0.0.1
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
Flask==3.0.2
